[FIX] delivery_multi_destination: Make delivery shipping properly with multi
Handle the step of sending the shipping (triggered on picking validation) for
multi destination delivery. This means to manually check for fixing prices (as
the implementation doesn't check this part and always take the main carrier price,
not the subcarrier one.

Test for this use case done, although the rest of the cases are not covered by
tests yet.